Fix issues and delete old code comments

+++ Aquarium/Aquarium/Camera.cpp
@@ -4,6 +4,7 @@
 #include <gtc/matrix_transform.hpp>
 #include <GL/glew.h>
 
+
 Camera::Camera(const int width, const int height, const glm::vec3& position)
 	: startPosition(position)
 {
@@ -17,66 +18,67 @@ glm::vec3 Camera::GetPosition()
 	return position;
 }
 
-
 void Camera::Set(const int width, const int height, const glm::vec3& position)
 {
-    this->isPerspective = true;
-    this->yaw = YAW;
-    this->pitch = PITCH;
+	this->isPerspective = true;
+	this->yaw = YAW;
+	this->pitch = PITCH;
 
-    this->FoVy = FOV;
-    this->width = width;
-    this->height = height;
-    this->zNear = zNEAR;
-    this->zFar = zFAR;
+	this->FoVy = FOV;
+	this->width = width;
+	this->height = height;
+	this->zNear = zNEAR;
+	this->zFar = zFAR;
 
-    this->worldUp = glm::vec3(0, 1, 0);
-    this->position = position;
+	this->worldUp = glm::vec3(0, 1, 0);
+	this->position = position;
 
-    lastX = width / 2.0f;
-    lastY = height / 2.0f;
-    bFirstMouseMove = true;
+	lastX = width / 2.0f;
+	lastY = height / 2.0f;
+	bFirstMouseMove = true;
 
-    UpdateCameraVectors();
+	UpdateCameraVectors();
 }
 
 void Camera::Reset(const int width, const int height)
 {
-    Set(width, height, startPosition);
+	Set(width, height, startPosition);
 }
 
 void Camera::Reshape(int windowWidth, int windowHeight)
 {
-    width = windowWidth;
-    height = windowHeight;
+	width = windowWidth;
+	height = windowHeight;
 
-    // define the viewport transformation
-    glViewport(0, 0, windowWidth, windowHeight);
+	// define the viewport transformation
+	glViewport(0, 0, windowWidth, windowHeight);
+
+	// Adaugare log pentru debug
 }
 
 const glm::mat4 Camera::GetViewMatrix() const
 {
-    // Returns the View Matrix
-    return glm::lookAt(position, position + forward, up);
+	// Returns the View Matrix
+	return glm::lookAt(position, position + forward, up);
 }
 
 const glm::mat4 Camera::GetProjectionMatrix() const
 {
-    glm::mat4 Proj = glm::mat4(1);
-    if (isPerspective) {
-        float aspectRatio = ((float)(width)) / height;
-        Proj = glm::perspective(glm::radians(FoVy), aspectRatio, zNear, zFar);
-    }
-    else {
-        float scaleFactor = 2000.f;
-        Proj = glm::ortho<float>(
-            -width / scaleFactor, width / scaleFactor,
-            -height / scaleFactor, height / scaleFactor, -zFar, zFar);
-    }
-    return Proj;
+	glm::mat4 Proj = glm::mat4(1);
+	if (isPerspective) {
+		float aspectRatio = ((float)(width)) / height;
+		Proj = glm::perspective(glm::radians(FoVy), aspectRatio, zNear, zFar);
+	}
+	else {
+		float scaleFactor = 2000.f;
+		Proj = glm::ortho<float>(
+			-width / scaleFactor, width / scaleFactor,
+			-height / scaleFactor, height / scaleFactor, -zFar, zFar);
+	}
+	return Proj;
 }
 
-void Camera::ProcessKeyboard(ECameraMovementType direction, float deltaTime)
+void Camera::ProcessKeyboard(ECameraMovementType direction, double deltaTime)
 {
 	float velocity = (float)(cameraSpeedFactor * deltaTime);
 	if (mode == FreeLook)
@@ -149,7 +151,7 @@ void Camera::MouseControl(float xPos, float yPos)
 	}
 	xChange *= mouseSensitivity;
 	yChange *= mouseSensitivity;
-
+	//if (mode == CameraMode::FreeLook) 
 	ProcessMouseMovement(xChange, yChange);
 }
 
@@ -168,7 +170,6 @@ void Camera::SetCameraMode(CameraMode mode)
 {
 	this->mode = mode;
 }
-
 CameraMode Camera::GetCameraMode()
 {
 	return this->mode;
@@ -184,12 +185,11 @@ glm::vec3 Camera::GetRotation()
 	return rotation;
 }
 
-void Camera::ProcessMouseMovement(float xOffset, float yOffset, bool constrainPitch)
+void Camera::ProcessMouseMovement(float xOffset, float yOffset, bool constrainPitch /*= true*/)
 {
 	yaw += xOffset;
 	pitch += yOffset;
 
-	
 	if (constrainPitch) {
 		if (pitch > 89.0f)
 			pitch = 89.0f;
